separate database for sql storage integration tests
